+++ src/components/LImageOverlay.vue
@@ -1,5 +1,5 @@
 <script>
-import { onMounted, ref, inject, nextTick } from "vue";
+import { onMounted, ref, inject, nextTick, markRaw } from "vue";
 import {
   remapEvents,
   propsBinder,
@@ -31,7 +31,9 @@ export default {
       const { imageOverlay, DomEvent } = useGlobalLeaflet
         ? WINDOW_OR_GLOBAL.L
         : await import("leaflet/dist/leaflet-src.esm");
-      leafletRef.value = imageOverlay(props.url, props.bounds, options);
+      leafletRef.value = markRaw(
+        imageOverlay(props.url, props.bounds, options)
+      );
 
       const listeners = remapEvents(context.attrs);
       DomEvent.on(leafletRef.value, listeners);