+++ src/components/LLayerGroup.vue
@@ -1,5 +1,5 @@
 <script>
-import { onMounted, ref, inject, nextTick } from "vue";
+import { onMounted, ref, inject, nextTick, markRaw } from "vue";
 import {
   remapEvents,
   propsBinder,
@@ -24,7 +24,7 @@ export default {
       const { layerGroup, DomEvent } = useGlobalLeaflet
         ? WINDOW_OR_GLOBAL.L
         : await import("leaflet/dist/leaflet-src.esm");
-      leafletRef.value = layerGroup(props.options);
+      leafletRef.value = markRaw(layerGroup(props.options));
 
       const listeners = remapEvents(context.attrs);
       DomEvent.on(leafletRef.value, listeners);