+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [#231 Possible Regression](https://github.com/vue-leaflet/vue-leaflet/issues/231).
 - Reworked attribution update handler to resolve
   [#165 Error when updating attribution prop for tileLayer](https://github.com/vue-leaflet/vue-leaflet/issues/165).
+- Default Leaflet icon is replaced by an empty div initially when a Marker has content in its slot that has not
+  yet rendered, resolving
+  [#170 Blinking default marker when using custom icon component](https://github.com/vue-leaflet/vue-leaflet/issues/170).
 
 
 ## [0.8.1] - 2023-01-29