+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Default Leaflet icon is replaced by an empty div initially when a Marker has content in its slot that has not
   yet rendered, resolving
   [#170 Blinking default marker when using custom icon component](https://github.com/vue-leaflet/vue-leaflet/issues/170).
+- Resolved [#138 vue3 this.$slots.default is not a function](https://github.com/vue-leaflet/vue-leaflet/issues/138).
 
 ### Changed